fix small error

+++ music/views.py
@@ -22,7 +22,7 @@ def upload(request):
         # This is the starting time of the first note calculated at the front-end
         start_time = float(request.POST['start_time'])
         # This is the selected BPM
-        bpm = int(request.POST['bpm'])
+        bpm = float(request.POST['bpm'])
         print("-------------start_time received: ", start_time)
         # Store the uploaded recording at /media/audios
         handle_recording(request.FILES['recording'])
